Run rustfmt & clippy

+++ src/stdweb.rs
@@ -7,7 +7,8 @@ use std_web::web::{window, TypedArray};
 use webgl_stdweb::{
     WebGL2RenderingContext, WebGLBuffer, WebGLFramebuffer, WebGLProgram, WebGLQuery,
     WebGLRenderbuffer, WebGLRenderingContext, WebGLSampler, WebGLShader, WebGLSync, WebGLTexture,
-    WebGLUniformLocation, WebGLVertexArrayObject, WebGLVertexArrayObjectOES, WebGLTransformFeedback
+    WebGLTransformFeedback, WebGLUniformLocation, WebGLVertexArrayObject,
+    WebGLVertexArrayObjectOES,
 };
 
 #[derive(Debug)]
@@ -376,12 +377,11 @@ impl HasContext for Context {
 
     unsafe fn delete_shader(&self, shader: Self::Shader) {
         let mut shaders = self.shaders.borrow_mut();
-        match shaders.1.remove(shader) {
-            Some(ref s) => match self.raw {
+        if let Some(ref s) = shaders.1.remove(shader) {
+            match self.raw {
                 RawRenderingContext::WebGl1(ref gl) => gl.delete_shader(Some(s)),
                 RawRenderingContext::WebGl2(ref gl) => gl.delete_shader(Some(s)),
-            },
-            None => {}
+            }
         }
     }
 
@@ -468,12 +468,11 @@ impl HasContext for Context {
 
     unsafe fn delete_program(&self, program: Self::Program) {
         let mut programs = self.programs.borrow_mut();
-        match programs.1.remove(program) {
-            Some(ref p) => match self.raw {
+        if let Some(ref p) = programs.1.remove(program) {
+            match self.raw {
                 RawRenderingContext::WebGl1(ref gl) => gl.delete_program(Some(p)),
                 RawRenderingContext::WebGl2(ref gl) => gl.delete_program(Some(p)),
-            },
-            None => {}
+            }
         }
     }
 
@@ -951,78 +950,71 @@ impl HasContext for Context {
 
     unsafe fn delete_buffer(&self, buffer: Self::Buffer) {
         let mut buffers = self.buffers.borrow_mut();
-        match buffers.1.remove(buffer) {
-            Some(ref b) => match self.raw {
+        if let Some(ref b) = buffers.1.remove(buffer) {
+            match self.raw {
                 RawRenderingContext::WebGl1(ref gl) => gl.delete_buffer(Some(b)),
                 RawRenderingContext::WebGl2(ref gl) => gl.delete_buffer(Some(b)),
-            },
-            None => {}
+            }
         }
     }
 
     unsafe fn delete_framebuffer(&self, framebuffer: Self::Framebuffer) {
         let mut framebuffers = self.framebuffers.borrow_mut();
-        match framebuffers.1.remove(framebuffer) {
-            Some(ref f) => match self.raw {
+        if let Some(ref f) = framebuffers.1.remove(framebuffer) {
+            match self.raw {
                 RawRenderingContext::WebGl1(ref gl) => gl.delete_framebuffer(Some(f)),
                 RawRenderingContext::WebGl2(ref gl) => gl.delete_framebuffer(Some(f)),
-            },
-            None => {}
+            }
         }
     }
 
     unsafe fn delete_query(&self, query: Self::Query) {
         let mut queries = self.queries.borrow_mut();
-        match queries.1.remove(query) {
-            Some(ref r) => match self.raw {
+        if let Some(ref r) = queries.1.remove(query) {
+            match self.raw {
                 RawRenderingContext::WebGl1(ref _gl) => panic!("Query objects are not supported"),
                 RawRenderingContext::WebGl2(ref gl) => gl.delete_query(Some(r)),
-            },
-            None => {}
+            }
         }
     }
 
     unsafe fn delete_renderbuffer(&self, renderbuffer: Self::Renderbuffer) {
         let mut renderbuffers = self.renderbuffers.borrow_mut();
-        match renderbuffers.1.remove(renderbuffer) {
-            Some(ref r) => match self.raw {
+        if let Some(ref r) = renderbuffers.1.remove(renderbuffer) {
+            match self.raw {
                 RawRenderingContext::WebGl1(ref gl) => gl.delete_renderbuffer(Some(r)),
                 RawRenderingContext::WebGl2(ref gl) => gl.delete_renderbuffer(Some(r)),
-            },
-            None => {}
+            }
         }
     }
 
     unsafe fn delete_sampler(&self, sampler: Self::Sampler) {
         let mut samplers = self.samplers.borrow_mut();
-        match samplers.1.remove(sampler) {
-            Some(ref s) => match self.raw {
+        if let Some(ref s) = samplers.1.remove(sampler) {
+            match self.raw {
                 RawRenderingContext::WebGl1(ref _gl) => panic!("Samplers are not supported"),
                 RawRenderingContext::WebGl2(ref gl) => gl.delete_sampler(Some(s)),
-            },
-            None => {}
+            }
         }
     }
 
     unsafe fn delete_sync(&self, fence: Self::Fence) {
         let mut fences = self.fences.borrow_mut();
-        match fences.1.remove(fence) {
-            Some(ref f) => match self.raw {
+        if let Some(ref f) = fences.1.remove(fence) {
+            match self.raw {
                 RawRenderingContext::WebGl1(ref _gl) => panic!("Fences are not supported"),
                 RawRenderingContext::WebGl2(ref gl) => gl.delete_sync(Some(f)),
-            },
-            None => {}
+            }
         }
     }
 
     unsafe fn delete_texture(&self, texture: Self::Texture) {
         let mut textures = self.textures.borrow_mut();
-        match textures.1.remove(texture) {
-            Some(ref t) => match self.raw {
+        if let Some(ref t) = textures.1.remove(texture) {
+            match self.raw {
                 RawRenderingContext::WebGl1(ref gl) => gl.delete_texture(Some(t)),
                 RawRenderingContext::WebGl2(ref gl) => gl.delete_texture(Some(t)),
-            },
-            None => {}
+            }
         }
     }
 
@@ -1374,7 +1366,7 @@ impl HasContext for Context {
         .try_into()
         // Errors will be caught by the browser or through `get_error`
         // so return a default instead
-        .unwrap_or(String::from(""))
+        .unwrap_or_else(|_| String::from(""))
     }
 
     unsafe fn get_parameter_string(&self, parameter: u32) -> String {
@@ -1385,7 +1377,7 @@ impl HasContext for Context {
         .try_into()
         // Errors will be caught by the browser or through `get_error`
         // so return a default instead
-        .unwrap_or(String::from(""))
+        .unwrap_or_else(|_| String::from(""))
     }
 
     unsafe fn get_uniform_location(
@@ -1723,7 +1715,13 @@ impl HasContext for Context {
         }
     }
 
-    unsafe fn tex_storage_1d(&self, _target: u32, _levels: i32, _internal_format: u32, _width: i32) {
+    unsafe fn tex_storage_1d(
+        &self,
+        _target: u32,
+        _levels: i32,
+        _internal_format: u32,
+        _width: i32,
+    ) {
         panic!("Tex storage 1D is not supported");
     }
 
@@ -2817,7 +2815,9 @@ impl HasContext for Context {
 
     unsafe fn create_transform_feedback(&self) -> Result<Self::TransformFeedback, String> {
         let raw_transform_feedback = match self.raw {
-            RawRenderingContext::WebGl1(ref _gl) => panic!("TransformFeedback objects are not supported"),
+            RawRenderingContext::WebGl1(ref _gl) => {
+                panic!("TransformFeedback objects are not supported")
+            }
             RawRenderingContext::WebGl2(ref gl) => gl.create_transform_feedback(),
         };
 
@@ -2833,74 +2833,106 @@ impl HasContext for Context {
 
     unsafe fn delete_transform_feedback(&self, transform_feedback: Self::TransformFeedback) {
         let mut transform_feedbacks = self.transform_feedbacks.borrow_mut();
-        match transform_feedbacks.1.remove(transform_feedback) {
-            Some(ref t) => match self.raw {
-                RawRenderingContext::WebGl1(ref _gl) =>  panic!("TransformFeedback objects are not supported"),
+        if let Some(ref t) = transform_feedbacks.1.remove(transform_feedback) {
+            match self.raw {
+                RawRenderingContext::WebGl1(ref _gl) => {
+                    panic!("TransformFeedback objects are not supported")
+                }
                 RawRenderingContext::WebGl2(ref gl) => gl.delete_transform_feedback(Some(t)),
-            },
-            None => {}
+            }
         }
     }
 
-    unsafe fn bind_transform_feedback(&self, target: u32, transform_feedback: Option<Self::TransformFeedback>) {
+    unsafe fn bind_transform_feedback(
+        &self,
+        target: u32,
+        transform_feedback: Option<Self::TransformFeedback>,
+    ) {
         let transform_feedbacks = self.transform_feedbacks.borrow();
-        let raw_transform_feedback = transform_feedback.map(|tf| transform_feedbacks.1.get_unchecked(tf));
+        let raw_transform_feedback =
+            transform_feedback.map(|tf| transform_feedbacks.1.get_unchecked(tf));
         match self.raw {
-            RawRenderingContext::WebGl1(ref _gl) => panic!("TransformFeedback objects are not supported"),
-            RawRenderingContext::WebGl2(ref gl) => gl.bind_transform_feedback(target, raw_transform_feedback)
+            RawRenderingContext::WebGl1(ref _gl) => {
+                panic!("TransformFeedback objects are not supported")
+            }
+            RawRenderingContext::WebGl2(ref gl) => {
+                gl.bind_transform_feedback(target, raw_transform_feedback)
+            }
         }
     }
 
     unsafe fn begin_transform_feedback(&self, primitive_mode: u32) {
         match self.raw {
-            RawRenderingContext::WebGl1(ref _gl) => panic!("TransformFeedback objects are not supported"),
-            RawRenderingContext::WebGl2(ref gl) => gl.begin_transform_feedback(primitive_mode)
+            RawRenderingContext::WebGl1(ref _gl) => {
+                panic!("TransformFeedback objects are not supported")
+            }
+            RawRenderingContext::WebGl2(ref gl) => gl.begin_transform_feedback(primitive_mode),
         }
     }
 
     unsafe fn end_transform_feedback(&self) {
         match self.raw {
-            RawRenderingContext::WebGl1(ref _gl) => panic!("TransformFeedback objects are not supported"),
-            RawRenderingContext::WebGl2(ref gl) => gl.end_transform_feedback()
+            RawRenderingContext::WebGl1(ref _gl) => {
+                panic!("TransformFeedback objects are not supported")
+            }
+            RawRenderingContext::WebGl2(ref gl) => gl.end_transform_feedback(),
         }
     }
 
     unsafe fn pause_transform_feedback(&self) {
         match self.raw {
-            RawRenderingContext::WebGl1(ref _gl) => panic!("TransformFeedback objects are not supported"),
-            RawRenderingContext::WebGl2(ref gl) => gl.pause_transform_feedback()
+            RawRenderingContext::WebGl1(ref _gl) => {
+                panic!("TransformFeedback objects are not supported")
+            }
+            RawRenderingContext::WebGl2(ref gl) => gl.pause_transform_feedback(),
         }
     }
 
     unsafe fn resume_transform_feedback(&self) {
         match self.raw {
-            RawRenderingContext::WebGl1(ref _gl) => panic!("TransformFeedback objects are not supported"),
-            RawRenderingContext::WebGl2(ref gl) => gl.resume_transform_feedback()
-        }
-    }
-
-    unsafe fn transform_feedback_varyings(&self, program: Self::Program, varyings: &[&str], buffer_mode: u32) {
-        let programs = self.programs.borrow();
-        let raw_program = programs.1.get_unchecked(program);
-        match self.raw {
-            RawRenderingContext::WebGl1(ref _gl) => panic!("TransformFeedback objects are not supported"),
-            RawRenderingContext::WebGl2(ref gl) => gl.transform_feedback_varyings(raw_program, varyings, buffer_mode),
-        }
-    }
-
-    unsafe fn get_transform_feedback_varying(&self, program: Self::Program, index: u32) -> Option<ActiveTransformFeedback> {
-        let programs = self.programs.borrow();
-        let raw_program = programs.1.get_unchecked(program);
-        match self.raw {
-            RawRenderingContext::WebGl1(ref _gl) => panic!("TransformFeedback objects are not supported"),
-            RawRenderingContext::WebGl2(ref gl) => {
-                gl.get_transform_feedback_varying(raw_program, index)
-                    .map(|info| ActiveTransformFeedback {
-                        size: info.size(),
-                        tftype: info.type_(),
-                        name: info.name(),
-                    })
+            RawRenderingContext::WebGl1(ref _gl) => {
+                panic!("TransformFeedback objects are not supported")
             }
+            RawRenderingContext::WebGl2(ref gl) => gl.resume_transform_feedback(),
+        }
+    }
+
+    unsafe fn transform_feedback_varyings(
+        &self,
+        program: Self::Program,
+        varyings: &[&str],
+        buffer_mode: u32,
+    ) {
+        let programs = self.programs.borrow();
+        let raw_program = programs.1.get_unchecked(program);
+        match self.raw {
+            RawRenderingContext::WebGl1(ref _gl) => {
+                panic!("TransformFeedback objects are not supported")
+            }
+            RawRenderingContext::WebGl2(ref gl) => {
+                gl.transform_feedback_varyings(raw_program, varyings, buffer_mode)
+            }
+        }
+    }
+
+    unsafe fn get_transform_feedback_varying(
+        &self,
+        program: Self::Program,
+        index: u32,
+    ) -> Option<ActiveTransformFeedback> {
+        let programs = self.programs.borrow();
+        let raw_program = programs.1.get_unchecked(program);
+        match self.raw {
+            RawRenderingContext::WebGl1(ref _gl) => {
+                panic!("TransformFeedback objects are not supported")
+            }
+            RawRenderingContext::WebGl2(ref gl) => gl
+                .get_transform_feedback_varying(raw_program, index)
+                .map(|info| ActiveTransformFeedback {
+                    size: info.size(),
+                    tftype: info.type_(),
+                    name: info.name(),
+                }),
         }
     }
 }
